collector/requirements.txt
pip
uuid
boto3
pyyaml
botocore
sqlalchemy
psycopg[binary]
pytenable
okta
azure-identity
azure-storage-blob
crowdstrike-falconpy
jira[cli]
dnspython
google-cloud-bigquery
dateparser
cloudflare==2.19.*
duckdb
duckdb-engine
python-dotenv
simple-salesforce
pandas
pandas_gbq
pip-system-certs
simple-salesforce